requirements.txt
asgiref==3.7.2
Django==4.2.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
